backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
pandas
python-dotenv
openai>=1.0.0,<2.0.0
